+++ android/app/test/screens/player/now_playing_lyrics_expand_test.dart
@@ -123,9 +123,11 @@ const String _syncedBody = '''
 void main() {
   setUp(() {
     paletteExtractorOverride = (Uri? _) async => null;
+    heroArtFloatEnabled = false;
   });
   tearDown(() {
     paletteExtractorOverride = dominantColorFor;
+    heroArtFloatEnabled = true;
   });
 
   testWidgets('lyrics render inside a card with an expand affordance',